debug comm errors for paper to user service. when a new user is registered, a score of 0 is automatically assigned

+++ SearchService/src/main/java/com/ynm/searchservice/dto/AuthorDto.java
@@ -12,6 +12,6 @@ import lombok.NoArgsConstructor;
 public class AuthorDto {
     private Integer id;
     private ResearchPaper paper;
-    private Integer userId;  // just the ID, not full User
+    private Long userId;  // just the ID, not full User
     private String position;
 }
